Default Eclipse to write to generated/java, not bin/generated/java
Despite the sanity of outputting generated source into the "default output directory", apparently Eclipse refuses to compile with that setting. Output to generated/java instead.

We used to output to generated_src, same as gradle build, but there are adverse side-effects when the two processes share the same output location.