Pass generated sources to FindBugs
We have always excluded generated sources from FindBugs verification, as users will not be able to fix any flagged issues. However, this causes issues when some validators need to find generated supertypes. This commit fixes this issue by packaging the generated sources up into a JAR, so the types are available but not validated.