Rely on gradle 4.6's annotationProcessor logic if that configuration exists (#83)
## Context

In gradle 4.6+, if there are any dependencies in the `annotationProcessor` configuration (added by default by the `java` plugin), then gradle automatically configures java compile tasks with `-processorpath` arguments.
There is actually one such configuration per sourceSet, so by default there is `annotationProcessor` (for `main`) and `testAnnotationProcessor` (for `test`).

## Before this PR

If the `annotationProcessor` configuration exists, we make our `processor` configuration extend from it. This conflicts with gradle's automatic handling of the `annotationProcessor` task, such that if that `annotationProcessor` is not empty, our logic (which was adding the relevant `-processorpath` flag to all JavaCompile tasks) gets superseded by gradle adding the same flag.

Therefore, dependencies in `processor` but not in `annotationProcessor` would not be picked up.

## After this PR

If the `annotationProcessor` configuration exists, this plugin no longer configures any `JavaCompile` tasks, and instead just makes all the annotationProcessor configurations for each source set extend our `processor` configuration, to ensure that processors added in older code are picked up and configured by gradle.

Additionally, we collect the `processor` AND (potentially) all `annotationProcessor` configurations into a sink configuration called `allProcessors`, and use that when configuring IDEs, to ensure that all processors will be included.